Dynamic transforms (#12)
* rename files

* remove entity registry class

* add uuid and scene classes

* separate file for mesh

* add scenegraph

* use scene graph

* partially done

* add animation class

* minor fixes

* cleanup

* use InstanceTagComponent

* add scene hierarchy panel

* use scene hierarchy panel

* pass scene to model loader

* misc changes

* fix scene hierarchy panel

* add instancebuffer class

* rename shaders

* implement instance buffer

* dynamic transform in gltf example

* fix viewport resize image deform issue

* fixes for gltf sample

* scene hierarchy panel in anim sample

* fix vertex normal issue

* fix compiler warnings

* use type size_t

* fix webgpu issues

+++ src/RenderSys/Vulkan/VulkanBuffer.cpp
@@ -3,7 +3,7 @@
 namespace RenderSys
 {
 
-VulkanBuffer::VulkanBuffer(uint32_t byteSize, RenderSys::BufferUsage bufferUsage)
+VulkanBuffer::VulkanBuffer(size_t byteSize, RenderSys::BufferUsage bufferUsage)
     : m_bufferSize(byteSize)
 {
     VkBufferCreateInfo bufferInfo{};
